requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.0.0
ultralytics>=8.0.0
# Use headless OpenCV to avoid GUI library dependencies (libGL.so.1)
# Note: If you still get libGL.so.1 errors, install system packages:
# Ubuntu/Debian: apt-get install -y libgl1-mesa-glx libglib2.0-0
# Alpine: apk add --no-cache libgl libglib
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
python-multipart>=0.0.6
easyocr>=1.7.0
google-generativeai>=0.3.0  # Legacy support
google-genai>=0.2.0  # New Gemini API
python-dotenv>=1.0.0
# Database dependencies
pymongo>=4.6.0
redis>=5.0.0

